requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv
Flask-Migrate==4.1.0
Pillow==10.4.0
requests==2.32.3
midtransclient==1.3.0